Initial proof of concept for docker proxy
For each incoming request, take a look at the host portion of the uri
request and slit it among '.'. For each sub domain, check to see if
there is a docker instance with that name. If you find one, get that
docker instance's public port and send the request there.